Add deterministic hook test for plugin ordering
- Created integration test file at crates/plugin/tests/hook_determinism.rs
- Tests verify hook execution is deterministic based on plugin priority
- Updated task P2-029-01 status to done

+++ opencode-rust/crates/plugin/tests/hook_determinism.rs
@@ -137,7 +137,7 @@ fn hook_order_is_deterministic() {
         fn on_start(&mut self) -> Result<(), PluginError> {
             let order = self.call_count.fetch_add(1, Ordering::SeqCst);
             let mut seq = self.call_sequence.lock().unwrap();
-            if seq.len() == order as usize {
+            if seq.len() == order {
                 seq.push(self.name.clone());
             } else {
                 seq.push(format!("OUT_OF_ORDER:{}", self.name));
@@ -217,7 +217,7 @@ fn hook_order_is_consistent_across_invocations() {
             fn on_start(&mut self) -> Result<(), PluginError> {
                 let order = self.call_count.fetch_add(1, Ordering::SeqCst);
                 let mut seq = self.call_sequence.lock().unwrap();
-                if seq.len() == order as usize {
+                if seq.len() == order {
                     seq.push(self.name.clone());
                 } else {
                     seq.push(format!("OUT_OF_ORDER:{}", self.name));
